+++ askbot/views/commands.py
@@ -612,12 +612,20 @@ def delete_tag(request):
     if request.user.is_anonymous() \
         or not request.user.is_administrator_or_moderator():
         raise exceptions.PermissionDenied()
-    post_data = simplejson.loads(request.raw_post_data)
-    tag_name = forms.clean_tag(post_data['tag_name'])
-    path = post_data['path']
-    tree = category_tree.get_data()
-    category_tree.delete_category(tree, tag_name, path)
-    category_tree.save_data(tree)
+
+    try:
+        post_data = simplejson.loads(request.raw_post_data)
+        tag_name = post_data['tag_name']
+        path = post_data['path']
+        tree = category_tree.get_data()
+        category_tree.delete_category(tree, tag_name, path)
+        category_tree.save_data(tree)
+    except Exception:
+        if 'tag_name' in locals():
+            logging.critical('could not delete tag %s' % tag_name)
+        else:
+            logging.critical('failed to parse post data %s' % request.raw_post_data)
+        raise exceptions.PermissionDenied(_('Sorry, could not delete tag'))
     return {'tree_data': tree}
 
 @csrf.csrf_exempt
@@ -691,7 +699,7 @@ def subscribe_for_tags(request):
             else:
                 message = _(
                     'Tag subscription was canceled (<a href="%(url)s">undo</a>).'
-                ) % {'url': request.path + '?tags=' + request.REQUEST['tags']}
+                ) % {'url': escape(request.path) + '?tags=' + request.REQUEST['tags']}
                 request.user.message_set.create(message = message)
             return HttpResponseRedirect(reverse('index'))
         else:
@@ -1481,7 +1489,7 @@ def get_editor(request):
     )
     #parse out javascript and dom, and return them separately
     #we need that, because js needs to be added in a special way
-    html_soup = BeautifulSoup(editor_html)
+    html_soup = BeautifulSoup(editor_html, 'html5lib')
 
     parsed_scripts = list()
     for script in html_soup.find_all('script'):