+++ askbot/views/writers.py
@@ -12,7 +12,6 @@ import random
 import sys
 import tempfile
 import time
-import urlparse
 from django.shortcuts import get_object_or_404
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
@@ -551,13 +550,6 @@ def edit_answer(request, id):
             if request.user.can_make_group_private_posts():
                 form.initial['post_privately'] = answer.is_private()
 
-        #gives a chance to set extra initial data on the form
-        signals.answer_before_editing.send(None,
-            answer=answer,
-            user=request.user,
-            form=form
-        )
-
         data = {
             'page_class': 'edit-answer-page',
             'active_tab': 'questions',
@@ -566,6 +558,13 @@ def edit_answer(request, id):
             'revision_form': revision_form,
             'form': form,
         }
+        extra_context = context.get_extra(
+            'ASKBOT_EDIT_ANSWER_PAGE_EXTRA_CONTEXT',
+            request,
+            data
+        )
+        data.update(extra_context)
+
         return render(request, 'answer_edit.html', data)
 
     except exceptions.PermissionDenied, e: